added server
local service discovery test server

local service discovery
